requirements.txt
python-telegram-bot
requests
whisper
TTS
ffmpeg-python
torch
sentencepiece